htmlGenerator/requirements.txt
jinja2
mistune
#ruamel.yaml
pyyaml
arrow
mistune-contrib
webassets
parse
cssmin
markupsafe
pygments
